refactorizacion carrito

+++ src/controllers/carts.js
@@ -1,30 +1,46 @@
 import { request, response } from "express"
-import { CartsRepository } from "../repositories/index.js"
+import { CartsRepository, ProductsRepository, UsersRepository } from "../repositories/index.js"
 
 export const getCartById = async(req=request, res=response) => {
     try {
+        const {_id}= req
         const {cid} = req.params
+        const usuario = await UsersRepository.getUserByID(_id)
+
+        if(!usuario) return res.status(400).json({ok:false,msg:'usuario no existe'})
+
+        if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok:false, msg:'carrito no valido'})
+
         const cart = await CartsRepository.getCartById(cid)
-        if(cart)
-           return res.json({cart})
-        return res.status(404).json({msg: `el carrito con id ${cid} no existe`})
+        return res.json({cart})
     } catch (error) {
         return response.status(500).json({msg: 'hablar con en administrados'})
     }
 }
 
-export const createCart = async(req=request, res=response) => {
+/*export const createCart = async(req=request, res=response) => {
     try {
         const cart = await CartsRepository.createCart()
         return res.json({msg: 'carrito creado', cart})
     } catch (error) {
         return response.status(500).json({msg: 'hablar con en administrados'})
     }
-}
+*/
 
 export const addProductInCart = async(req=request, res=response) => {
     try {
+        const {_id}= req
         const {cid, pid} =  req.params
+
+        const usuario = await UsersRepository.getUserByID(_id)
+
+        if(!usuario) return res.status(400).json({ok:false,msg:'usuario no existe'})
+
+        if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok:false, msg:'carrito no valido'})
+
+        const existeProducto = await ProductsRepository.getProductById(pid)
+        if(!existeProducto) return res.status(400).json({ok:false, msg:'el producto no existe'})
+
         const cart = await CartsRepository.addProductInCart(cid, pid)
         if(!cart)
             return res.status(404).json({msg:`el carrito conid ${cid} no existe`})
@@ -37,10 +53,18 @@ export const addProductInCart = async(req=request, res=response) => {
 
 export const deleteProductsInCart= async (req = request, res = response)=>{
     try {
+        const {_id}= req
         const {cid, pid} = req.params
+
+        const usuario = await UsersRepository.getUserByID(_id)
+
+        if(!usuario) return res.status(400).json({ok:false,msg:'usuario no existe'})
+        if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok:false, msg:'carrito no valido'})
+
+        const existeProducto = await ProductsRepository.getProductById(pid)
+        if(!existeProducto) return res.status(400).json({ok:false, msg:'el producto no existe'})
+
         const cart = await CartsRepository.deleteProductsInCart(cid,pid)
-        if(!cart)
-           return res.status(404).json({msg:'no se pudo relizar la operacion'})
         return res.json({msg:'producto eliminado del carrito', cart})
     } catch (error) {
         return res.status(500).json({msg: 'hablar con el admin'})
@@ -50,8 +74,18 @@ export const deleteProductsInCart= async (req = request, res = response)=>{
 
 export const updateProductsInCart= async (req = request, res = response)=>{
     try {
+        const {_id}= req
         const {cid, pid} = req.params
         const {quantity} = req.body
+
+        const usuario = await UsersRepository.getUserByID(_id)
+
+        if(!usuario) return res.status(400).json({ok:false,msg:'usuario no existe'})
+        if(!(usuario.cart_id.toString() === cid)) return res.status(400).json({ok:false, msg:'carrito no valido'})
+
+        const existeProducto = await ProductsRepository.getProductById(pid)
+        if(!existeProducto) return res.status(400).json({ok:false, msg:'el producto no existe'})
+
 
         if(!quantity || Number.isInteger(quantity))
            return res.status(404).json({msg:'quantity es obligatorio y debe ser un numero entero'})
@@ -66,7 +100,7 @@ export const updateProductsInCart= async (req = request, res = response)=>{
     }
 }
 
-export const deleteCart= async (req = request, res = response)=>{
+/*export const deleteCart= async (req = request, res = response)=>{
     try {
         const {cid} = req.params
         const cart = await CartsRepository.deleteCart(cid)
@@ -78,3 +112,4 @@ export const deleteCart= async (req = request, res = response)=>{
         
     }
 }
+*/
